Refactor members middleware out of router
