+++ routes/index.js
@@ -1,12 +1,14 @@
 const express = require('express')
 const router = express.Router()
 
-const Member = require('./members/model')
+const members = require('./members/middleware')
 
 router.get('/', (req, res) => {
   res.send({
     title: 'High Skill Masters API',
   })
 })
+
+router.post('/subscribe', members.subscribe)
 
 module.exports = router
